+++ examples/src/CustomBunnymarkExample/CustomBunnymark.js
@@ -6,6 +6,7 @@ import Particle from "./Particle";
 import bunnys from "./bunnys.png";
 
 const maxSize = 200000;
+const initialBunnies = 100;
 const bunniesAddedPerFrame = 100;
 const gravity = 0.5;
 const maxX = 800;
@@ -69,20 +70,16 @@ class CustomBunnymark extends Component {
     const bunny5 = new PIXI.Texture(bunnyTextures.baseTexture, new PIXI.Rectangle(2, 2, 26, 37));
 
     this.bunnyTextures = [bunny1, bunny2, bunny3, bunny4, bunny5];
-    const currentTexture = 2;
     this.setState({
-      bunnys: [generateBunny(currentTexture), generateBunny(currentTexture)],
-      currentTexture: currentTexture,
+      bunnys: Array(initialBunnies)
+        .fill(0)
+        .map(() => generateBunny(Math.floor(Math.random() * this.bunnyTextures.length))),
     });
 
     this.props.app.ticker.add(this.animate);
   }
 
-  componentWillUnmount() {
-    this.props.app.ticker.remove(this.animate);
-  }
-
-  animate = () => {
+  componentDidUpdate(prevProps, prevState, snapshot) {
     const { bunnys, currentTexture, isAdding } = this.state;
 
     if (isAdding) {
@@ -95,9 +92,18 @@ class CustomBunnymark extends Component {
       }
       const newBunnys = bunnys.concat(addedBunnys);
 
-      this.setState({ bunnys: newBunnys });
+      // add more bunnies in next frame
+      requestAnimationFrame(() => {
+        this.setState({ bunnys: newBunnys });
+      });
     }
+  }
 
+  componentWillUnmount() {
+    this.props.app.ticker.remove(this.animate);
+  }
+
+  animate = () => {
     this.particleContainer.children.forEach(bunny => bunny.update(bunny));
   };