+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Added support for React Concurrent Mode
+
 
 ## [1.0.0-beta.1] - 2020-01-19